+++ src/components/spec-document/try-it/TryIt.vue
@@ -10,7 +10,9 @@
         size="small"
       />
       <span class="path">{{ data.path }}</span>
-      <TryItButton
+
+      <TryItDropdown
+        class="tryit-dropdown"
         :data="data"
         @tryit-api-call="doApiCall"
       />
@@ -59,7 +61,7 @@
 <script setup lang="ts">
 import { inject, computed, ref, watch } from 'vue'
 import type { PropType, Ref } from 'vue'
-import TryItButton from './TryItButton.vue'
+import TryItDropdown from './TryItDropdown.vue'
 import { getRequestHeaders } from '@/utils'
 import type { IHttpOperation } from '@stoplight/types'
 import MethodBadge from '@/components/common/MethodBadge.vue'
@@ -211,6 +213,9 @@ watch(() => (props.data.id), () => {
     padding: var(--kui-space-50, $kui-space-50);
     .path {
       margin-left: var(--kui-space-20, $kui-space-20);
+    }
+    .tryit-dropdown {
+      margin-left: auto;
     }
   }