+++ src/types/response-code-dot.ts
@@ -1,0 +1,5 @@
+export const ResponseCodeDotVariants = ['2xx', '4xx'] as const
+export type ResponseCode = typeof ResponseCodeDotVariants[number]
+
+export const ResponseCodeDotSizeVariants = ['small', 'large'] as const
+export type ResponseCodeDotSize = typeof ResponseCodeDotSizeVariants[number]